refactor: 음식점 정보/추가 관련 css 모듈 리팩토링 및 오타 수정
