+++ src/components/asides/AddRestaurantModal.jsx
@@ -1,4 +1,4 @@
-import styles from "./RestaurantModal.model.css";
+import styles from "./RestaurantModal.module.css";
 
 function AddRestaurantModal() {
   return (
